requirements.txt
-i https://pypi.org/simple
aiobotocore==2.4.1
aiohttp==3.8.3 ; python_version >= '3.6'
aioitertools==0.11.0 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
amqp==5.1.1 ; python_version >= '3.6'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==22.1.0 ; python_version >= '3.5'
awscrt==0.14.0
billiard==3.6.4.0
boto3[crt]==1.24.59 ; python_version >= '3.7'
botocore==1.27.59 ; python_version >= '3.7'
celery[sqs]==5.2.7
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
click==8.1.3 ; python_version >= '3.7'
click-didyoumean==0.3.0 ; python_full_version >= '3.6.2' and python_full_version < '4.0.0'
click-plugins==1.1.1
click-repl==0.2.0
coloredlogs==15.0.1
frozenlist==1.3.3 ; python_version >= '3.7'
humanfriendly==10.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.4 ; python_version >= '3.5'
jmespath==1.0.1 ; python_version >= '3.7'
kombu==5.2.4 ; python_version >= '3.7'
multidict==6.0.3 ; python_version >= '3.7'
prompt-toolkit==3.0.36 ; python_full_version >= '3.6.2'
pycurl==7.44.1
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1 ; python_version >= '3.7' and python_version < '4'
s3transfer==0.6.0 ; python_version >= '3.7'
-e .
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.4.0 ; python_version < '3.10'
urllib3==1.26.13 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
vine==5.0.0 ; python_version >= '3.6'
wcwidth==0.2.5
wrapt==1.14.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
yarl==1.8.2 ; python_version >= '3.7'
